moved swift tempalte implementation to the framework
