+++ Sourcery/Generating/Generator.swift
@@ -9,6 +9,6 @@ import SourceryRuntime
 enum Generator {
     static func generate(_ types: Types, template: Template, arguments: [String: NSObject] = [:]) throws -> String {
         Log.verbose("Rendering template \(template.sourcePath)")
-        return try template.render(types: types, arguments: arguments)
+        return try template.render(TemplateContext(types: types, arguments: arguments))
     }
 }
