+++ Sourcery/Generating/Template/Template.swift
@@ -6,7 +6,6 @@
 import Foundation
 import PathKit
 import SourceryRuntime
-import SourceryJS
 
 /// Generic template that can be used for any of the Sourcery output variants
 protocol Template {
@@ -19,5 +18,5 @@ protocol Template {
     /// - Parameter arguments: List of template arguments.
     /// - Returns: Generated code.
     /// - Throws: `Throws` template errors
-    func render(types: Types, arguments: [String: NSObject]) throws -> String
+    func render(_ context: TemplateContext) throws -> String
 }
