+++ Sourcery/Sourcery.swift
@@ -8,6 +8,7 @@ import PathKit
 import SwiftTryCatch
 import SourceryRuntime
 import SourceryJS
+import SourcerySwift
 
 class Sourcery {
     public static let version: String = inUnitTests ? "Major.Minor.Patch" : Version.current.value
