+++ SourceryTests/Generating/SwiftTemplateSpecs.swift
@@ -12,6 +12,7 @@ import Nimble
 import PathKit
 @testable import Sourcery
 @testable import SourceryRuntime
+@testable import SourcerySwift
 
 class SwiftTemplateTests: QuickSpec {
     override func spec() {
@@ -98,8 +99,8 @@ class SwiftTemplateTests: QuickSpec {
                     try Generator.generate(Types(types: []), template: SwiftTemplate(path: templatePath))
                     }
                     .to(throwError(closure: { (error) in
-                        let path = Path.cleanTemporaryDir(name: "build") + "main.swift"
-                        expect("\(error)").to(equal("\(path):6:13: error: expected expression in list of expressions\n  print(\"\\( )\", terminator: \"\");\n            ^\n"))
+                        let path = Path.cleanTemporaryDir(name: "build").parent() + "SwiftTemplate.build/main.swift"
+                        expect("\(error)").to(equal("\(path):6:19: error: expected expression in list of expressions\n        print(\"\\( )\", terminator: \"\");\n                  ^\n"))
                     }))
             }
